// src/types/country.ts
export interface Country {
  name: Name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  cioc?: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  subregion?: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  borders?: string[];
  area: number;
  demonyms?: Demonyms;
  flag: string;
  maps: Maps;
  population: number;
  gini?: Gini;
  fifa?: string;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  postalCode?: PostalCode;
}

export interface Name {
  common: string;
  official: string;
  nativeName?: NativeName;
}

export interface NativeName {
  fra?: Fra;
  ara?: Ara;
  spa?: Spa;
  jpn?: Jpn;
  nor?: Nor;
  arc?: Arc;
  ckb?: Ckb;
  eng?: Eng;
  nya?: Nya;
  kat?: Kat;
  ita?: Ita;
  vie?: Vie;
  est?: Est;
  kon?: Kon;
  lin?: Lin;
  lua?: Lua;
  swa?: Swa;
  ind?: Ind;
  div?: Div;
  por?: Por;
  zho?: Zho;
  mfe?: Mfe;
  srp?: Srp;
  tet?: Tet;
  afr?: Afr;
  deu?: Deu;
  her?: Her;
  hgm?: Hgm;
  kwn?: Kwn;
  loz?: Loz;
  ndo?: Ndo;
  tsn?: Tsn;
  rus?: Rus;
  tgk?: Tgk;
  isl?: Isl;
  hun?: Hun;
  msa?: Msa;
  tam?: Tam;
  bjz?: Bjz;
  zdj?: Zdj;
  mri?: Mri;
  nzs?: Nzs;
  ben?: Ben;
  amh?: Amh;
  kor?: Kor;
  nld?: Nld;
  pap?: Pap;
  urd?: Urd;
  crs?: Crs;
  sot?: Sot;
  cat?: Cat;
  jam?: Jam;
  gil?: Gil;
  cha?: Cha;
  lao?: Lao;
  rar?: Rar;
  pih?: Pih;
  heb?: Heb;
  prs?: Prs;
  pus?: Pus;
  tuk?: Tuk;
  hat?: Hat;
  glv?: Glv;
  mah?: Mah;
  run?: Run;
  lat?: Lat;
  ell?: Ell;
  ton?: Ton;
  tha?: Tha;
  slv?: Slv;
  dan?: Dan;
  fao?: Fao;
  ltz?: Ltz;
  smo?: Smo;
  gsw?: Gsw;
  roh?: Roh;
  pol?: Pol;
  ssw?: Ssw;
  bel?: Bel;
  mya?: Mya;
  mon?: Mon;
  hin?: Hin;
  swe?: Swe;
  aym?: Aym;
  que?: Que;
  hmo?: Hmo;
  tpi?: Tpi;
  kal?: Kal;
  cnr?: Cnr;
  hye?: Hye;
  kir?: Kir;
  nrf?: Nrf;
  bwg?: Bwg;
  kck?: Kck;
  khi?: Khi;
  ndc?: Ndc;
  nde?: Nde;
  sna?: Sna;
  toi?: Toi;
  tso?: Tso;
  ven?: Ven;
  xho?: Xho;
  zib?: Zib;
  khm?: Khm;
  ron?: Ron;
  grn?: Grn;
  bis?: Bis;
  fil?: Fil;
  tur?: Tur;
  pov?: Pov;
  lit?: Lit;
  slk?: Slk;
  tir?: Tir;
  fas?: Fas;
  bos?: Bos;
  hrv?: Hrv;
  fin?: Fin;
  ukr?: Ukr;
  mkd?: Mkd;
  uzb?: Uzb;
  ces?: Ces;
  mlt?: Mlt;
  ber?: Ber;
  aze?: Aze;
  dzo?: Dzo;
  nau?: Nau;
  sqi?: Sqi;
  niu?: Niu;
  pau?: Pau;
  som?: Som;
  mlg?: Mlg;
  mey?: Mey;
  gle?: Gle;
  sin?: Sin;
  nbl?: Nbl;
  nso?: Nso;
  zul?: Zul;
  cal?: Cal;
  kin?: Kin;
  tvl?: Tvl;
  nep?: Nep;
  bar?: Bar;
  nfr?: Nfr;
  sag?: Sag;
  fij?: Fij;
  hif?: Hif;
  nno?: Nno;
  nob?: Nob;
  smi?: Smi;
  lav?: Lav;
  kaz?: Kaz;
  bul?: Bul;
  tkl?: Tkl;
}
// ...
export interface Currencies {
  XAF?: Xaf;
  QAR?: Qar;
  DOP?: Dop;
  JPY?: Jpy;
  XOF?: Xof;
  NOK?: Nok;
  IQD?: Iqd;
  MWK?: Mwk;
  GEL?: Gel;
  USD?: Usd;
  GNF?: Gnf;
  EUR?: Eur;
  VND?: Vnd;
  CDF?: Cdf;
  IDR?: Idr;
  MVR?: Mvr;
  XPF?: Xpf;
  MOP?: Mop;
  MUR?: Mur;
  RSD?: Rsd;
  EGP?: Egp;
  NAD?: Nad;
  ZAR?: Zar;
  TJS?: Tjs;
  ISK?: Isk;
  HUF?: Huf;
  CHF?: Chf;
  SGD?: Sgd;
  UGX?: Ugx;
  BZD?: Bzd;
  KMF?: Kmf;
  NZD?: Nzd;
  BDT?: Bdt;
  ETB?: Etb;
  KRW?: Krw;
  ANG?: Ang;
  SAR?: Sar;
  PKR?: Pkr;
  BMD?: Bmd;
  AOA?: Aoa;
  SCR?: Scr;
  GHS?: Ghs;
  XCD?: Xcd;
  AUD?: Aud;
  LSL?: Lsl;
  CLP?: Clp;
  ILS?: Ils;
  JOD?: Jod;
  JMD?: Jmd;
  KID?: Kid;
  GBP?: Gbp;
  SHP?: Shp;
  LAK?: Lak;
  CKD?: Ckd;
  NIO?: Nio;
  UYU?: Uyu;
  LYD?: Lyd;
  AFN?: Afn;
  HTG?: Htg;
  IMP?: Imp;
  BHD?: Bhd;
  BIF?: Bif;
  TOP?: Top;
  BSD?: Bsd;
  THB?: Thb;
  DKK?: Dkk;
  FOK?: Fok;
  WST?: Wst;
  PAB?: Pab;
  VES?: Ves;
  PLN?: Pln;
  SZL?: Szl;
  BYN?: Byn;
  KYD?: Kyd;
  KPW?: Kpw;
  RUB?: Rub;
  MMK?: Mmk;
  COP?: Cop;
  CNY?: Cny;
  CRC?: Crc;
  CAD?: Cad;
  TZS?: Tzs;
  MNT?: Mnt;
  TWD?: Twd;
  INR?: Inr;
  OMR?: Omr;
  SDG?: Sdg;
  SEK?: Sek;
  PEN?: Pen;
  TTD?: Ttd;
  ZMW?: Zmw;
  PGK?: Pgk;
  LBP?: Lbp;
  AMD?: Amd;
  GYD?: Gyd;
  KGS?: Kgs;
  JEP?: Jep;
  ZWL?: Zwl;
  KHR?: Khr;
  RON?: Ron2;
  BOB?: Bob;
  VUV?: Vuv;
  PHP?: Php;
  TRY?: Try;
  ERN?: Ern;
  CUC?: Cuc;
  CUP?: Cup;
  IRR?: Irr;
  BAM?: Bam;
  UAH?: Uah;
  SBD?: Sbd;
  SRD?: Srd;
  MKD?: Mkd2;
  HKD?: Hkd;
  TND?: Tnd;
  UZS?: Uzs;
  BWP?: Bwp;
  SSP?: Ssp;
  AED?: Aed;
  AWG?: Awg;
  GIP?: Gip;
  CZK?: Czk;
  KES?: Kes;
  GTQ?: Gtq;
  MAD?: Mad;
  MXN?: Mxn;
  AZN?: Azn;
  ARS?: Ars;
  BTN?: Btn;
  BND?: Bnd;
  STN?: Stn;
  BBD?: Bbd;
  BGN?: Bgn;
  TMT?: Tmt;
  KZT?: Kzt;
  FKP?: Fkp;
  FJD?: Fjd;
  DJF?: Djf;
  MRU?: Mru;
  GGP?: Ggp;
  MDL?: Mdl;
  MZN?: Mzn;
  BRL?: Brl;
  CVE?: Cve;
  ALL?: All;
  NGN?: Ngn;
  GMD?: Gmd;
  SOS?: Sos;
  YER?: Yer;
  MYR?: Myr;
  MGA?: Mga;
  DZD?: Dzd;
  PYG?: Pyg;
  LKR?: Lkr;
  KWD?: Kwd;
  SLL?: Sll;
  RWF?: Rwf;
  SYP?: Syp;
  HNL?: Hnl;
  TVD?: Tvd;
  NPR?: Npr;
  LRD?: Lrd;
}

export interface Xaf {
  name: string;
  symbol: string;
}

export interface Qar {
  name: string;
  symbol: string;
}

export interface Dop {
  name: string;
  symbol: string;
}

export interface Jpy {
  name: string;
  symbol: string;
}

export interface Xof {
  name: string;
  symbol: string;
}

export interface Nok {
  name: string;
  symbol: string;
}

export interface Iqd {
  name: string;
  symbol: string;
}

export interface Mwk {
  name: string;
  symbol: string;
}

export interface Gel {
  name: string;
  symbol: string;
}

export interface Usd {
  name: string;
  symbol: string;
}

export interface Gnf {
  name: string;
  symbol: string;
}

export interface Eur {
  name: string;
  symbol: string;
}

export interface Vnd {
  name: string;
  symbol: string;
}

export interface Cdf {
  name: string;
  symbol: string;
}

export interface Idr {
  name: string;
  symbol: string;
}

export interface Mvr {
  name: string;
  symbol: string;
}

export interface Xpf {
  name: string;
  symbol: string;
}

export interface Mop {
  name: string;
  symbol: string;
}

export interface Mur {
  name: string;
  symbol: string;
}

export interface Rsd {
  name: string;
  symbol: string;
}

export interface Egp {
  name: string;
  symbol: string;
}

export interface Nad {
  name: string;
  symbol: string;
}

export interface Zar {
  name: string;
  symbol: string;
}

export interface Tjs {
  name: string;
  symbol: string;
}

export interface Isk {
  name: string;
  symbol: string;
}

export interface Huf {
  name: string;
  symbol: string;
}

export interface Chf {
  name: string;
  symbol: string;
}

export interface Sgd {
  name: string;
  symbol: string;
}

export interface Ugx {
  name: string;
  symbol: string;
}

export interface Bzd {
  name: string;
  symbol: string;
}

export interface Kmf {
  name: string;
  symbol: string;
}

export interface Nzd {
  name: string;
  symbol: string;
}

export interface Bdt {
  name: string;
  symbol: string;
}

export interface Etb {
  name: string;
  symbol: string;
}

export interface Krw {
  name: string;
  symbol: string;
}

export interface Ang {
  name: string;
  symbol: string;
}

export interface Sar {
  name: string;
  symbol: string;
}

export interface Pkr {
  name: string;
  symbol: string;
}

export interface Bmd {
  name: string;
  symbol: string;
}

export interface Aoa {
  name: string;
  symbol: string;
}

export interface Scr {
  name: string;
  symbol: string;
}

export interface Ghs {
  name: string;
  symbol: string;
}

export interface Xcd {
  name: string;
  symbol: string;
}

export interface Aud {
  name: string;
  symbol: string;
}

export interface Lsl {
  name: string;
  symbol: string;
}

export interface Clp {
  name: string;
  symbol: string;
}

export interface Ils {
  name: string;
  symbol: string;
}

export interface Jod {
  name: string;
  symbol: string;
}

export interface Jmd {
  name: string;
  symbol: string;
}

export interface Kid {
  name: string;
  symbol: string;
}

export interface Gbp {
  name: string;
  symbol: string;
}

export interface Shp {
  name: string;
  symbol: string;
}

export interface Lak {
  name: string;
  symbol: string;
}

export interface Ckd {
  name: string;
  symbol: string;
}

export interface Nio {
  name: string;
  symbol: string;
}

export interface Uyu {
  name: string;
  symbol: string;
}
// ...
export interface Idd {
  root?: string;
  suffixes?: string[];
}

export interface Languages {
  fra?: string;
  ara?: string;
  spa?: string;
  jpn?: string;
  nor?: string;
  arc?: string;
  ckb?: string;
  eng?: string;
  nya?: string;
  kat?: string;
  ita?: string;
  vie?: string;
  est?: string;
  kon?: string;
  lin?: string;
  lua?: string;
  swa?: string;
  ind?: string;
  div?: string;
  por?: string;
  zho?: string;
  mfe?: string;
  srp?: string;
  tet?: string;
  afr?: string;
  deu?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  tsn?: string;
  rus?: string;
  tgk?: string;
  isl?: string;
  hun?: string;
  msa?: string;
  tam?: string;
  bjz?: string;
  zdj?: string;
  mri?: string;
  nzs?: string;
  ben?: string;
  amh?: string;
  kor?: string;
  nld?: string;
  pap?: string;
  urd?: string;
  crs?: string;
  sot?: string;
  cat?: string;
  jam?: string;
  eus?: string;
  glc?: string;
  gil?: string;
  cha?: string;
  lao?: string;
  rar?: string;
  pih?: string;
  heb?: string;
  prs?: string;
  pus?: string;
  tuk?: string;
  hat?: string;
  glv?: string;
  mah?: string;
  run?: string;
  lat?: string;
  ell?: string;
  ton?: string;
  tha?: string;
  slv?: string;
  dan?: string;
  fao?: string;
  ltz?: string;
  smo?: string;
  gsw?: string;
  roh?: string;
  pol?: string;
  ssw?: string;
  bel?: string;
  mya?: string;
  mon?: string;
  hin?: string;
  swe?: string;
  aym?: string;
  que?: string;
  hmo?: string;
  tpi?: string;
  kal?: string;
  cnr?: string;
  hye?: string;
  kir?: string;
  nrf?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  sna?: string;
  toi?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zib?: string;
  khm?: string;
  ron?: string;
  grn?: string;
  bis?: string;
  fil?: string;
  tur?: string;
  pov?: string;
  lit?: string;
  slk?: string;
  tir?: string;
  fas?: string;
  bos?: string;
  hrv?: string;
  fin?: string;
  ukr?: string;
  mkd?: string;
  uzb?: string;
  ces?: string;
  mlt?: string;
  ber?: string;
  aze?: string;
  dzo?: string;
  nau?: string;
  sqi?: string;
  niu?: string;
  pau?: string;
  som?: string;
  mlg?: string;
  mey?: string;
  gle?: string;
  sin?: string;
  nbl?: string;
  nso?: string;
  zul?: string;
  cal?: string;
  kin?: string;
  tvl?: string;
  nep?: string;
  de?: string;
  nfr?: string;
  sag?: string;
  fij?: string;
  hif?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  lav?: string;
  kaz?: string;
  bul?: string;
  tkl?: string;
}

export interface Translations {
  ara: Ara2;
  bre: Bre;
  ces: Ces2;
  cym: Cym;
  deu: Deu2;
  est: Est2;
  fin: Fin2;
  fra: Fra2;
  hrv?: Hrv2;
  hun: Hun2;
  ita: Ita2;
  jpn?: Jpn2;
  kor: Kor2;
  nld: Nld2;
  per?: Per;
  pol: Pol2;
  por: Por2;
  rus: Rus2;
  slk: Slk2;
  spa: Spa2;
  srp: Srp2;
  swe: Swe2;
  tur: Tur2;
  urd: Urd2;
  zho?: Zho2;
}
// ...
export interface Demonyms {
  eng: Eng2;
  fra?: Fra3;
}

export interface Eng2 {
  f: string;
  m: string;
}

export interface Fra3 {
  f: string;
  m: string;
}

export interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

export interface Gini {
  "2017"?: number;
  "2019"?: number;
  "2013"?: number;
  "2015"?: number;
  "2012"?: number;
  "2016"?: number;
  "2011"?: number;
  "2018"?: number;
  "2014"?: number;
  "1999"?: number;
  "2009"?: number;
  "2004"?: number;
  "2006"?: number;
  "1992"?: number;
  "1998"?: number;
  "2010"?: number;
  "2003"?: number;
  "2005"?: number;
  "2008"?: number;
}

export interface Car {
  signs?: string[];
  side: string;
}

export interface Flags {
  png: string;
  svg: string;
  alt?: string;
}

export interface CoatOfArms {
  png?: string;
  svg?: string;
}

export interface CapitalInfo {
  latlng?: number[];
}

export interface PostalCode {
  format: string;
  regex?: string;
}